package.json, sea-site-build - facilitate link to playground for developer convenience
Developing sea-map needs a way to deploy a mini map website. Doing
this via a dependency from (say) the `playground` project to `sea-map`
is a bit cumbersome, because changes to sea-map need to be propagated
back via a push to github.  Changing playground's dependency on
sea-map to a relative path also doesn't work well, because a) this
change mustn't be checked in, but more seriously b) it doesn't then
pull in the requirejs dependency.

A simpler way is to symlink to playground within sea-map, using a link
called (say) `ext`, and adding an npm script task to build and serve
that directly.

See documentation added to README.md