app.js - parse the url params, allow them to override config/attributes
URL parameters preserve case, so don't need snake-case conversion.

They do permit the same parameter several times, so we need to flatten
those.  We discard all but the last.

This now permits config parameters to be overridden by URL parameters,
or HTML attributes, when there is a setter defined. (Currently there
are no setters defined, in the committed code).