app/model/sse_initiative.js - exclude complete list from registeredValues
i.e. the All category: put it in a separate variable. This way we
can't confuse it with something which happens to have the label 'All',
and the registeredValues structure is uniform.

Currently this means the All category isn't exposed in
registeredValues, but we'll fix this in an upcoming commit.