invert logic/rename useCache option to noLodCache
We want the default to be to enable the cache. And this to apply when
no parameter present, too. Calling it useCache suggests that when
absent, the cache *isn't* used.

Also: be more strict about the allowed values. Only allow "true" or
"false" (and don't interpret "false" as true!). Simplify logic where
possible.

Update and expand the documentation somewhat.