map-app/app/model/sse_initiative.js - fix/simplify reset method
And changeDatasets in map-app/app/presenter/sidebar/datasets.js

The logic is a bit wrong. The initial check in changeDatasets rejects
pretty much any change as "no change". And the two parameters just
make it harder to follow.

The double argument choice seems a bit gratuituous, considering the
choice of the internal value of currentDatasets. Why not just reflect
that choice?  So we do, and fix the logic at the same time, making the
reset method have the same behaviour.