bin/sea-site-deploy - add this deploy script
It encapsulates the rather lengthy rsync command which was inline in
the package.jon runscript.

It can be used both in sea-map, and in consuming map applications.

Additionally, I've implemented command-line options which can override
the configured values. This is for scripting, specifically the `seodo`
script (defined outside this repository currently).

See documentation inlined in the script.