bin/sea-site-build - port to use webpack
Otherwise, aim to completely stay a drop-in replacement for the old version.

Build css file to load statically in index.html - emulates sea-map
behaviour of yore... allows consumers to supply their own index.html.

Don't shell-out to `date`. Use Date.toISOString instead - which
admittedly always returns UTC time, so no timezone, but I think that's
only a minor difference we probably don't care about.

We keep the bulk of the JS in bin/sea-site-build.js, and just require
it from a script in the old file path. This is for the convenience of
my editor (emacs' web-mode), which seems to be foxed by nodejs
scripts which don't end in .js.

bin/sea-site-build.js etc. - implement support for optional config/popup.js

Bonus: it can stay an AMD module! Webpack seems to detect that
automatically. We do need to define an alias so that the
sse_initiatives import works.

Remove webpack.config.js, so these don't have to be both maintained.