map-app/app/view/map/marker.js - avoid errors when markers not in active dataset
Seems that in multi-dataset mode, when a single dataset is selected,
initiatives from the other datasets can be listed in the directory
still, and clicking them tries to focus a non-existant pin.

Avoid this causing exceptions - although this may require a deeper fix
because possibly the initiatives should be removed from the directory.