app/model/sse_initiative.js - make loadDataset() async
i.e. Make it return the result wrapped in a promise, rather than
trying to resolve the promise itself. Put the common code in handler
functions.

This is a first step to enabling the operation to be chained to vocab
data loads (which will be async too).  And possibly later to
parallelising the dispatch...