app/model/sse_initiative.js - simplify asSearchStr initialisation
Integrate the special field information into the classSchema and use
that instead.

Note that we no longer de-dupliate the (odd) cases where a field has a
singular and multiple form (with a plural name). These cases may
result in duplicate values in the search string, but because of the
nature of search, this doesn't matter. Whereas, it vastly simplifies
the logic required - just one pass through the searchedFields looking
up matches in classSchema to deduce what to do in each case.