app/model/sse_initiative.js - use the `init` attribute to indicate multi-value properties
Not the presence of `oldStyleKey` as before - this was a mistake, they
just happen to co-incide because of a historical assumption.

We can generalise later, for now we're just changing the logic of
selecting fields to accumulate from several rows by the initialiser,
which is more closely indicating a multi-value "type" when it is
`asList`.