app/model/sse_initiative.js - look up searchedFields better
Previously it was inserting the corresponding term URI into the
searchable string when the property was initialised `asList`. This is
because there was no check for `isList`, and no `oldStyleKey`
attributes for `isList` properties. So add these.

It's still not quite perfect, you can't list properties which don't
have a key in the oldStyleValues object, or which don't have a value
in the parameters object (like 'activities').  We'll fix the former
later when we replace oldStyleValues with a direct look-up in the
vocabs index. The latter will mean that field cannot be searched,
because there is no such binding the query, and will be fixable by
adjusting the config / query.