app/model/sse_initiative.js - replace oldStyleKey labels with vocabUris
This has the consequence that we don't store vocabs in initiative
fields as just the ID, but as an abbreviated URI, which will match the
vocabs index. What we're aiming for is to get rid of the
oldStyleValues lookup, at least internally to sse_initiative.js.

This is good, but also has a knock on effect elsewhere in the code,
where the form of these IDs are assumed to be what they were.

So also fix app/view/sidebar/directory.js not to expect these codes to
be only alphanumeric IDs. Don't try and uppercase them when matching,
etc.

Making this fix work was rather painful, but I think it does.