app/view/sidebar/initiatives.js - repair selection of searched terms
In the search sidebar, select a value in a drop-down. The selection
works, but the selected value is lost - all drop downs revert to
--Any--.

This was again due to IDs being compared against labels, I think the
result of the content of some fields changing (like Country) from a
raw string to a vocab Uri.