app/model/sse_initiative.js - expect abbreviation to the max
Because get_vocabs now does not insist we have at least one character after the
colon. So this:

   http://dev.lod.coop/essglobal/2.1/vocabs/standard/regions-ica/

Used to be abbreviated to:

   essglobal:regions-ica/

Instead, it now allows it to match the whole URL with a prefix. So it
goes to just:

   reg:

(Because it matches the prefix reg:)

So update our URIs accordingly. Later we shall refactor this entirely
into the config.