app/model/sse_initiative.js - sort registeredValues by label, not ID
This was sorting by vocab ID, and resulting in a
not-totally-alphabetical sort from the viewpoint of the
labels. e.g. You get United Arab Emirates sorting to the top of the
country directory, because its country code is AE.

Sorting by lables requires doing a kind of Schwartzian transform to
get the labels to sort with. Luckily we have a vocab index to use for
mapping IDs to labels.