app/model/sse_initiative.js - add a `shortPostcode` field
This is derived from the postcode field. It strips out the first part
of the postcode, if one is present.

It's somewhat of a specific hack for the Owned-by-Oxford site, no
other site currently needs this... but I plan to strip down the fields
in here and allow external site-specific configuration, so these
fields will mostly all go soon, I hope (including this one).