model/config_schema.js - more thorough validation of lang code
We need to normalise and validate, so that we have upper case language
codes, a consistent fallback code, and no funny characters are
allowed (which might cause havoc if inserted in HTML blindly).

And do this here rather than in sse_initiative.js, so that it is
shared by all consumers (i.e. not just sse_initiative.js).  If we
don't, then language codes seen by (for example) the about.js code
will differ from that in the model.