presenter/sidebar.js - fix defaultOpenSidebar logic
When false, it still seems to open, because the Initiative.reset and
Initiative.complete events trigger p.changeSidebar() which then
triggers p.showSidebar().

Stop changeSidebar() from triggering showSidebar(), and alter the
logic so that all cases which do still need the sidebar shown call it
explicitly.  And those that don't, now don't trigger the sidebar to
open.

This seems to now work correctly for both cases of the
defaultOpenSidebar option, hopefully I've not broken some less obvious
case!