tsconfig.json - workaround to allow ts-node to execute scripts
Workaround found here:

https://stackoverflow.com/a/63459149/2960236

Essentially, without this ts-node / node get very confused when using
typescript scripts which include modules from elsewhere. There are
conflicting warnings, where either the import module syntax isn't
understood, or the import is interpreted as Javascript.

The post says there is a bug in ts-node, which is still open and
unsolved, but that this workaround suffices in the mean time.

https://github.com/TypeStrong/ts-node/issues/1007