bin/sea-site-build, etc. - enable site-specific entry point
i.e. ts/index.ts or ts/index.js

At least in the dev mode for now!

We also need to:

Adjust app.ts to be typescript friendlier (complaints about app.ts not
being a module)

Disable site-local popup.js support (we'll allow sites can import it
into the config in later commits).

Add an optional sea-site-build parameter -s to define where
site-specific source is kept.  Use this to find any index.js or
index.ts entry points. Otherwise use app.ts.

Add `"exclude": [ "./ext/" ]` to tsconfig.json in order that
typescript files in ./ext are not loaded by `tsc` unless explicitly
linked via an entry point or an include. (This allows the case where
the default `app.ts` is used as an entry point.)

And in order for this to work with index.ts (not index.js), but not
part of this commit, the symlinked project under ext/ needs a
tsconfig.json with a path linking back to the map-app source dir:

    {
      "compilerOptions": {
	"paths": {
	  "sea-map/*": ["../www/map-app/*"]
	},
	[...]
      }
    }