model/sse_initiative.ts - refactor Initiative.nongeo away
This appears to be a flag set if the initiative initially had no
location, but gets put in the default location.

Refactor the check for location into a function (which won't appear in
the properties list of an initiative), and create the marker at the
default location directly, without first setting the lat/lng
coordinates to the default.