app/model/sparqldataloader.ts - pre-process multi-record instances into single ones
So this:

    [{ uri: 'xxx', multivalue: 1 },
     { uri: 'xxx', multivalue: 2 },
     { uri: 'yyy', multivalue: 3 },
     { uri: 'zzz', multivalue: 4 }]

Becomes this:

    [{ uri: 'xxx', multivalue: [1,2] },
     { uri: 'yyy', multivalue: 3 },
     { uri: 'zzz', multivalue: 4 }]

Note that yyy and zzz's multivalue fields are represented by a single
non-array value. This is due to the lack of any metadata at the time
of pre-processing.

The SparqlDataAggregator then turns these objects into Initiatives in
one step. This makes it simpler to switch to a source which is not
coming from sparql (like a CSV).