app/model/dataservices.ts - make manual lat/lng interpretation less weird
lat/lng can be overridden by manLat/manLng.  But the check is for a coercing
equality with the literal string "0", which is an odd choice, and
means that "0" and 0 both match, but null or undefined don't, meaning
that undefined manLat/manLng overrides lat/lng!

Adjust the logic to allow numeric values (inferred by
Number.parseFloat returning non-NaN).  This solves the immediate
problem.