app/model/dataservices.ts - replace isAlpha in mkLocFromParam with something sane
isAlpha looks for alphabetic characters, and for some reason the logic
for parsing lat/lng values uses this to exclude non-numeric
values. However it doesn't work on actual floats! So "" passes, "1.1"
passes, but 1.1 fails...

Our data is now sending lat/lngs as numbers not strings, so obviously
this doesn't work any more.  Use something which parses a string or a
number correctly - Number.parseFloat.