package.json - work around instant-mocha's peer-dep on vulnerable mocha@8
Having installed mocha@10 using `npm install mocha@10` to work around
the audit report of vulnerabilities in mocha@8's dependencies, and
finding that the tests still work... See
https://github.com/privatenumber/instant-mocha/issues/34

I find that npm install now bails because of the peer dependency
instant-mocha@1.4.1 has on `mocha@"^6.1.4 || ^8.3.2"`.

Report made here:

https://github.com/privatenumber/instant-mocha/issues/38

However, issue #34 has a workaround, an override for this peer
dependency, which works with npm 8 and above (which we are using).

This commit applies this workaround.  It should be removed later, when
instant-mocha resolves this problem.