app/model/dataservices.ts - teach Initiative how to make its own constructor
Basically, here we refactor the code which dynamically creates
Initiative objects with certain fields defined by PropDefs, and values
defined by ParamBuilder<PropDef>, into a factory method of Initiative.

This method creates a constructor for Initiative instances which given
an InitiativeObj, will return an Initiative with the desired
properties and values.

This isn't very TypeScript-ish, but we have to do it at runtime
because the definitions are only read at runtime.

It could probably be further improved, but this is a step forwards -
the constructor code is now inside Initiative!  And creating and
testing Initiatives is now possible.