map-app/* - remove getAllRegisteredValues method and allRegisteredValues prop
At this point, whilst testing the code, we discover that
allRegisteredValues and its accessor method are never used any more.

This makes sense because it's an object mapping property titles to all
Initiatives with that property - which is all initiatives. So we have
a bunch of Initiative arrays containing all Initiatives. It doesn't
seem very useful!

So remove it!