app/model/propertyindexer.ts - implement multi-value, and vocab field support
So we should now properly index:
- `vocab` properties
- `multi` properties (of `value` or `vocab`, or `custom`)

Custom properties should work much like value properties, so long as
they return single values. (Multi-valued custom fields should be
implemented as multi-of-custom fields.)

Also, attempt to spot problems in the parameters up front in the
constructor. Specifically, spot clobbering of entries in
`propIdByVocabUri`. (We'd like to spot invalid vocabs, but this isn't
possible with the getVocab callback alone, we have to wait until a bad
URL is hit.)