app/presenter/sidebar/initiatives.js - refactor with comments
This is a key use of the vocabFilteredFields index created during
aggregation. I'm investigating it as it cannot support two properties
in the initiative schema using the same vocab - because it is keyed by
vocab ID, not property ID.

I think it needs to be inverted, but to do that I need to investigate
the uses of this index, such as here.