app/model/vocabs.ts - eliminate vocabFilteredFields usage from #getTerms
getTerms doesn't actually need to use vocabFilteredFields, it can get
it's information if it orders its algorithm slightly differently and
iterates over propertySchema looking for vocabs there.

We are aiming to eliminate vocabFilteredFields, which is a hash
mapping (abbreviated) vocab URIs to property ids. Mainly because, as
this is keyed by vocab URI, this hash cannot support the case where
more than one property uses the same vocab.

And if we eliminate vocabFilteredFields, also there is one less thing to
build when aggregating the data.