app/presenter/sidebar/base.js - de-weird the contentStack/stackWidth properties
It was, for whatever reason, set up as properties on the
class *prototype*, like the methods.  This means, like the methods,
all derived classes inherit it as a property.

This is a bit unusual - and I'm not sure if it was intended or
not. Possibly the author at the time got tangled up in JS prototype
class definitions, which are admittedly confusing.  But the
consequence is that each sub-class shares these properties,
read-write, which means they all have the same width (makes sense?)
and content stack (might make sense?)

To preserve this behaviour when converting to ES, I picked the
simplest analogue - to initialise these in a static class constructor.
It works. But it doesn't really play nicely with TypeScript, which is
the next step we want to get to.

Therefore, in this commit, I refactor this arrangement so that these
properties are on the parent sidebar presenter, and are accessed by
the child presenters via a parent property.

It's a bit ugly because of the parallel class heirarchies in use, but
I want to take small baby steps here - if we can get TypeScript
enabled, we can refactor with more confidence.