app/view/sidebar/base.js - de-weird the has*/title properties
These were, for whatever reason, set up as properties on the
class *prototype*, like the methods.  This means, like the methods,
all derived classes inherit it as a property.

This is a bit unusual - and I'm not sure if it was intended or
not. Possibly the author at the time got tangled up in JS prototype
class definitions, which are admittedly confusing.  But the
consequence is that each sub-class shares these properties,
read-write, which means they all have the same width (makes sense?)
and content stack (might make sense?)

To preserve this behaviour when converting to ES, I picked the
simplest analogue - to initialise these in a static class constructor.
It works. But it doesn't really play nicely with TypeScript, which is
the next step we want to get to.

Therefore, in this commit, I refactor this arrangement so that these
properties are normal propertie - inspecting the code in derived
classes, they appear to have been intended to be overridden.