app/view/sidebar/initiatives.js - typescriptify
And en route, typescriptify app/presenter.ts,
app/presenter/sidebar/base.ts, and move the SearchResults class into
its own typescript file.

This is because the contentStack of the parent BaseSidebarPresenter
class would be untyped otherwise, and InitiativesSidebarPresenter
needs it to be typed.

As a bonus, eliminate one usage of AggregatedData#vocabFilteredFields,
which is one of our end goals.