app/view/sidebar.js - typescriptify
We needed to refactor some of the HTML wrangling code to use d3
methods more uniformly where possible, as the typechecking seems to
get a bit hung up on the DOM methods' lack of typing. Specifically
insertBefore was problematic - replaced that with d3.insert. Also,
some of the transition event handlers seemed to have problems - since
there seemed not to be a similar solution there, I needed to add
several coercions.

And of course, there are quite a lot of cases where potentially
undefined values could be inserted where there should be strings -
guards added for these.