app/model/propertyindexer.ts - remove propIdByVocabUri property
And the corresponding vocabFilteredFields property from
DataAggregator.

We've refactored the code not to need this any more. This index
inherently does not permit more than one Initiative property to use a
given vocab, which is a bug - many properties might use the same
vocab.

There is more to fix, but this is a major step to removing this
restriction.