app/* - adjust the view/presenter class heirarchy: drop the generics
The generics add a problem with recursion.  So class A<B> needs to
link to class B<A>, but the type parameters don't stop there, it's an
infinite regress: A<B<A<B<A<B.... >>>>>

So don't do that. Instead, drop the presenter/view property from the
base classes, make them and the non-leaf classes abstract, and put
view/presenter properties back concretely in the leaf classes, or in
non-leaf classes as abstract properties, where required.

Note, this seems to have flushed out some more type errors, but we
don't fix those here.