app/*/map.js - typescriptify
Whew, this spiralled into a bit of a mission.  The map view/controller
are really quite complicated, unusually encoupled to everything else,
includes a number of nasty hacks (like putting an instance of mykoMap
on the window object??) and the Leaflet plug-ins are not all
supporting typescript.

I don't think this fully works currently - the centering seems to be broken.