app/*/map/marker.js - typescriptify
Code doesn't quite compile as some things fail the typechecking.
However, it's hard to know if that's because they are not working or
just that they're unorthodox / hacky ways of achieving some goal that
works. So for now I've left them in, and will remove them in the next
commit - with the intention that this makes these changes more clear
to other devs / future versions of me, if this needs to be addressed
later.

One case which failed typechecks, but which I worked around, was the
unofficial Marker property hasPhysicalLocation being used as a boolean
flag. In this case I created the ExtendedMarker interface and coercced
the marker type to that where it was needed. Hopefully this leaves a
clear trail in the code as well, and can be addressed later when there
are better tests / understanding of what's going on.