app/*/map/marker.js - remove small sections I can't typescriptify
Code wasn't quite compiling as these things fail the typechecking.
However, it's hard to know if that's because they are not working or
just that they're unorthodox / hacky ways of achieving some goal that
works. I'm remove them under the assumption that it's the former, but
with the intention that this makes these changes more clear to other
devs / future versions of me, if this needs to be addressed later.

The one case is the use of the cluster flag in the options passed to
the leaflet.AwesomeMarkers.icon method - this looks like it has
nothing to do with AwesomeMarkers, presumably it might have something
to do with MarkerClusterGroups, but I can't find any mention of it in
the documentation for the latter.

The other is the inclusion of an initiative property to the options
for the same method.  The looks like a hack, as none of the leaflet
plugins would have any concept of an initiative as used here.
However, I can't see any evidence that the information stashed here is
extracted and used from the options anywhere else.