app/*/map.ts - remove what is commented a "fix for firefox"
As this causes an instability in the map pins when selecting markers -
after the map has zoomed and panned, in some cases the pop-up gets
dismissed, and in some cases it does not.  When it is, this seems to
be accompanied by a glitchy map redreaw, along with all the markers.

I'm tentatively removing the map .refresh method and eventbus event as
these seem only to serve this fix... If problems arise, however, I may
need to revert this.