app/presenter/sidebar/datasets.ts - invert view/presenter relationship
Instead of view creating presenter, make presenter create view.  This
is a pattern we'll be implementing through the UI class heirarchy.

We can also simplify the parameters by passing a mapui, or access it
via the presenter in the case of the view.

We also fix the sidebar refresh mechanism - possible to do simply only
now that all the child sidebars have been converted to presenters
rather than views.