app/presenter/map/marker.ts - invert view/presenter relationship
Instead of view creating presenter, make presenter create view.  This
is a pattern we'll be implementing through the UI class heirarchy.

We can also simplify the parameters by passing a mapui, or access it
via the presenter in the case of the view.

MApUI also needs some reconfiguration to create the right things and
supply the right parameters.