app/model/config_schema.ts - add directoryField config parameter
We need to be able to set non-vocab properties to index on the
directory, filterableFields used to be the way to do this (first item
is the directory) but now it strictly only accepts vocab fields.

This config allows these two to be decoupled.  We can still allow the
default to be the first of filterableFields.