app/map-ui.ts - first pass at refactoring out a FilterService class
This is an attempt to make the filtering functions stand-alone and
testable.

This step wasn't particularly easy, and I needed to alter the
FilterService.filter data structure to be a Dictionary of Dictionaries
of I, in order to avoid needing to index the .uri property of the
Initiative class.  Reason: I want FilterService not to depend on the
things being filtered, other testing it gets fiddly.

Otherwise, things are just renamed somewhat, to avoid the cumbersomely
long names used before.