app/map-ui.ts - use Ids instead of Initiative more in FilterService
In FilterService, .hidden can be just a list of IDs, the full
initiatives aren't really used. So make them so.

This requires changing MarkerManager.showMarkers() and .hideMarkers()
to accept URI lists, which is a simple change.

Add FilterService.getFilteredIds() and .getUnfilteredIds() to
facilitate the transparent use of IDs rather than initiatives.

Convert lists of initatives to lists of IDs where appropriate.