app/map-ui.ts - use sets of I in FilterService
This gets a bit of the best of both worlds - object comparisons are
fast, probably faster than string comparisons, and the class is still
stand-alone as it can be parameterised to use anything.

Unfortunately we need to undo some of the work in earlier commits to
use URIs instead of initiatives. Not all of it though.