app/marker-manager.ts - use a Map<Initiative, MapMarkerPresenter> internally
Instead of a Dictionary<MapMarkerPresenter>. This means we can
object-compare the keys, and get other benefits of JS Maps.  Although
iteration of Maps is a bit more of a PITA in plain JS/TS.

This will prove very convenient when we change the way the seach state
is managed shortly.