app/map-ui.ts - make MapFilter.verboseName non-optional
This seems to have only been optional to allow the string searches and
the field filter to share the same data structure.  Having split the
former into MapSearch and the later into MapFilter, it's no longer
necessary to have it optional, and this simplifies a lot of defensive
code.

Plus, we are going to want it to be non-optional when we move
integrate the .verboseNamesMap field of the FilterService into
.filters