app/map-ui.ts,search-results.ts etc. - be explicit about recording the vocab title and terms
The previous implementation concatenated them together like `${title}:
${term}`, and then split them apart when needed. This obscures what is
going on, and makes the whole process harder to unpick.

Let's be explicit - put these in dedicated properties for the task.
Then we can see what is going where, and we will not be not stuck in
situations where we need the title and can't get it.  We don't need to
split the string to get these values any more.

(Not to mention that we shouldn't be indexing our datastructures on
localised terms, but we can now attempt to fix that, later.)