app/view/sidebar/initiatives.ts - wire filter dropdown ids for propName...
They were previously wired for the vocab title - which obvs means only
one vocab can be shown, excluding the possibility for multiple fields
with the same vocab.

We can't quite get rid of that yet, because propDefs have no
localisation - we only have the symbolic name for property!

We now also create one drop-down per filterableFields entry.  This
means some vocabs may be duplicated.  For now - don't do that! But we
don't protect you from it. Yet!  We should, or implement support for
multipe fields with the same vocab.

This refactor is a step towards givng the filters propName and
propValue fields, which in turn opens the door to a better filter
implementation waiting in the wings.