app/map-ui.ts - add .propName and .propValue to MapFilter
This forces all the filtering machinery to say which property name and
value are being filtered on.

It's a bit complicated at the moment, because filters are still
partially per-vocab, although now there is one dropdown per
property.

But this at least exposes which property we are filtering
on, which we need to wire up a better filtering mechanism waiting in
the wings.