app/presenter/map.ts - fix mm #176
MapPresenter.applyFilter() checks for filters, and if it thinks there
are none, calls .removeFilters(). Unclear why - the logic is a bit
loopy to follow - but in this case, the check is wrong - it calls
.length on an object which isn't an array, so gets 0 even if it isn't
empty.

The correct thing to do is call the right method which gets something
easy to get the emptiness of (without explicitly calling
Object.keys). .getFilters() does that already and returns an array.

This appears to fix the bug where selecting an item in the drop-down,
then changing to another, resets the drop-down to "-Any-"