app/model/data-loader.ts - move .search() to Initiative.textSearch()
This seems to be a sensible home for it, if it accepts a list of
initiatives to search.  This is better, as it allows arbitrary lists
to be searched, as opposed to just the current list of all of them
stored in a DataLoader.

Also rename sortInitiatives() function as Initiative.compare()