app/{view,presenter}/sidebar/initiatives.ts - push logic for search back as far as poss. wip
i.e. Don't check for empty strings where the search is called, and
duplicate that everywhere, with a switch to calling
.performSearchNoText(). The logic is broken, and inconsistent. Not all
filters get a search. Worse, .changeFilters re-applies a text search
and resets all the initiatives to hidden.

This is a work in progress, not everything is deduplicated and fixed.