app/marker-manager.ts - replace .hideMarkers and .showMarkers with .updateVisibility
I reckon having two methods is probably overthinking it, trying to
optimise the update when there is no proven bottleneck here (not
within the last three years at least!)

Having just one method makes the logic simpler, and therefore easier
to refactor, which is the task in hand.

Note, I will be using a set to keep track of the filtered initiatives
later, so I anticipate that here by accepting a set as the
argument. This means converting arrays to sets, but this will
disappear later when this other change is implemented.