app/view/sidebar/base.ts - make .title abstract, and localise subclass' .title
This fixes a bug (noted, but not in an issue) where the search panel
is titled "Untitled", which is the default set in the base sidebar
view class.

Make the base class field abstract - this forces all subclasses to
implement it explicitly.

Also, the other subclasses implement it already, but not in a
localised way. Localise them all too - the phrases are in the
phrasebook already.