app/view/sidebar/directory.ts,utils.ts - be less hardwired looking up allCountries
A partial fix for https://github.com/DigitalCommons/mykomap/issues/177

This is still a hack. We check the vocab URI to see if it looks like
one of our country-iso vocab URIs - which means we are no longer
checking a fixed set of languages with translations which might not
even match the ones in use, if they change.

This is good. However, it only works on vocab fields. - However I
don't think we currently use any country fields which aren't for the
directory.  This hack is mostly here to support ICA and DotCoop.
Possibly they should just make do with a generic "All Entries" title?