app/view/map.ts - hide location-less initiative markers
They get the default location, and their pop-up should appear when
selected, but they should not have a marker.

Fixed this with a bit of trial and error - it really isn't obvious to
me what's going on.  The variable naming seems to be a bit of a
red-herring too. selectedClusterGroup appears to be a group into which
locationless initiatives go, and unselectedClusterGroup is where the
ones which *do* go.  They both cluster - or did.

The big clue was that removing the addLayer() call on the latter makes
it disappear, so that only the locationless markers appear. And
removing it for the former leaves the locationless markers
hidden. But... also no popup.

So we disableClusteringAtZoom as zoomed out as possible.  This appears
to do the trick... markers gone, popups still pop up.

Fixes https://github.com/digitalcommons/mykomap/issues/192