package.json - update @types/d3 to ^7.4.0
Experiencing problems compiling related to typechecks on the return
type of select not matching d3Selection defined in this codebase.

Somewhat complicated and I don't fully understand the error, but as
the MykoMap code compiled ok with the ObO project, figured it was a
dependency version problem.

And this, although it needs a complete wipe of node_modules and
package-lock.json, seems to do the trick.