app/view/sidebar/initiatives.ts- submit text search on both blur and submit events
Currently, with it just on submit, we can get the situation described
here:

https://github.com/digitalcommons/mykomap/issues/197

i.e. alter the search text, and don't submit but instead alter a
filter - and the search text is reset.

This fix works by submitting as soon as the text input focus is lost,
so the reset happens but resets to the newly searched text.