app/model/data-services.ts - tease apart InnerDef from PropDef
InnerDefs can be nested inside a MultiPropDef. PropDefs can't.

This removes ambiguity - does the `from` property sit in the nested
part or the outer part of a MultiPropDef?

Unfortunately we picked the other scheme earlier - `from` is in the
nested part, but this doesn't work now we've teased those definitions
apart.  So adjust it. We'll need a breaking release.