app/model/property-indexer.ts - index by propName, not by title
This means we can find items by propName, we don't need title.

And it means we can fix a bug, in which the filterableFields wasn't
being used to set the directory index, instead the first key of
registeredValues was picked (incorrect!). It can be fixed because we
can get the right directory index from filterableFields, which wasn't
possible if the registeredValues array was indexed by title.

It also seems to fix this bug in which directory searches don't show
any pins (that must be related).

https://github.com/DigitalCommons/mykomap/issues/220

Also, title isn't actually guaranteed unique (although it wouldn't
make sense for two schema properties to have the same title)