package-lock.json - npm install, and bump to lockfileVersion 3
I'm not sure why this is happening now, as I've been using node
v18.18.0 for some time I believe, which comes with npm v9.8.1.

https://docs.npmjs.com/cli/v7/configuring-npm/package-lock-json

Says:

 - No version provided: an "ancient" shrinkwrap file from a version of npm prior to npm v5.
 - 1: The lockfile version used by npm v5 and v6.
 - 2: The lockfile version used by npm v7, which is backwards compatible to v1 lockfiles.
 - 3: The lockfile version used by npm v7, without backwards compatibility affordances. This is used for the hidden lockfile at node_modules/.package-lock.json, and will likely be used in a future version of npm, once support for npm v6 is no longer relevant.

Which implies we should be getting v3 already. But I guess it's fine
to use v3 so long as we do it everywhere. The package.json file
mandates npm >= 8.15.0 currently, so that should be checked and
enforced by npm?