app/model/csv-data-loader.ts - allow blank lines in CSVs
The the parser sees an EOL on the final line as denoting an invalid
blank line. When in fact that's kind of normal for CSVs, at least on
Unix. And so we will get an error on every CSV we parse with such a
final EOL. We probably care about spurious blank lines less than
spurious errors like this, so let's use this option, which seems to be
the only (simple?) workaround.

Discussed here: https://github.com/mholt/PapaParse/issues/150