app/{model/config-schema.ts,presenter/sidebar.ts} - make defaultPanel typesafe
Use typescipt's typing to restrict the possible values of defaultPanel
to one of the allowed ones.

Ran into a problem triggering ERR_REQUIRE_ESM with the d3 library,
just by changing the imports... turns out this is a thorny thing to
solve, so avoided it by not importing the things. Instead I needed to
list the allowed panel IDs in the documentation manually, rather than
programatically.

But ESM modules are "the future" so we probably need to convert
Mykomap to support it (and support CommonJS for cases that need that,
probably using Babel transpilation)