app/model/config-schema.ts - revise documentation for config
It was still talking about config.json, which is no longer being used,
hasn't been for some time. Describe how the current config is set via
a function parameter, and attempt to provide a potted example.

(Part of the conclusion I make is the potted example is forced to be a
bit more complicated than I think is ideal. )