app/model/config-schema.ts - add {get,set}FilteredPropDefs()
The new methods gather filterable property information from the
`fields` property (and uses "property" rather than "field" for
consistency with the more recent attempt to use JS-idiomatic terms.)
However they aim to be backward compatible if filterableFields is used
- the behaviour should be as before, but with the addition that any
filter default set in the new way is respected, if present.

Mark getFilterableFields() as deprecated.

Add validation of filterableFields in setFilterableFields() and in
constructor, remove validation elsewhere in DataServicesImpl.